fix debug hash function

+++ src/core/debug.h
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <iomanip>
 #define DEBUG(expr) std::cerr << "DEBUG " << expr << std::endl
-#define HEX(x) std::setw(2) << std::setfill('0') << std::hex << (int)(x) << std::setw(0) << std::dec
+#define HEX(x) std::setw(2) << std::setfill('0') << std::hex << (uint64_t)(x) << std::setw(0) << std::dec
 #else
 #define DEBUG(expr)
 #endif
